@@ -5,32 +5,32 @@ homepage: https://alphacephei.com/
 repository: https://github.com/alphacep/vosk-flutter
 
 environment:
-  sdk: ">=2.15.1 <3.0.0"
-  flutter: ">=2.5.0"
+  sdk: ">=3.5.2 <4.0.0"
+  flutter: ">=3.24.0"
 
 dependencies:
-  archive: ^3.1.8
-  args: ^2.4.0
+  archive: ^3.6.1
+  args: ^2.5.0
   build_cli_annotations: ^2.1.0
-  ffi: ^2.0.1
+  ffi: ^2.1.3
   flutter:
     sdk: flutter
-  http: ^1.2.0
+  http: ^1.2.2
   package_config: ^2.1.0
-  path: ^1.8.2
-  path_provider: ^2.0.8
-  permission_handler: ^10.2.0
-  pubspec_parse: ^1.2.2
+  path: ^1.9.0
+  path_provider: ^2.1.4
+  permission_handler: ^11.3.1
+  pubspec_parse: ^1.3.0
 
 dev_dependencies:
-  build_cli: ^2.2.1
-  build_runner: ^2.3.3
-  ffigen: ^7.2.9
+  build_cli: ^2.2.3
+  build_runner: ^2.4.12
+  ffigen: ^14.0.1
   flutter_test:
     sdk: flutter
-  mockito: ^5.3.2
-  mocktail: ^0.3.0
-  very_good_analysis: ^4.0.0+1
+  mockito: ^5.4.4
+  mocktail: ^1.0.4
+  very_good_analysis: ^6.0.0
 
 ffigen:
   name: VoskLibrary
